fix: Switching to Long id representation and global exception handling with enum entity types.

+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/controller/QuizController.java
@@ -20,7 +20,7 @@ public class QuizController {
   }
 
   @GetMapping("/{id}") //localhost:8080/quiz/1
-  public String quizModelView(@PathVariable("id") int id, Model model) {
+  public String quizModelView(@PathVariable("id") Long id, Model model) {
     Quiz quiz = this.quizService.findById(id);
     model.addAttribute("quiz", quiz);
     return "quiz";
