+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/service/ObjectService.java
@@ -4,6 +4,8 @@ import java.util.List;
 
 public interface ObjectService<T> {
   public void save(T t);
-  public T findById(int id);
+
+  public T findById(Long id);
+
   public List<T> findAll();
 }
