+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/service/QuizService.java
@@ -22,7 +22,7 @@ public class QuizService implements ObjectService<Quiz> {
   }
 
   @Override
-  public Quiz findById(int id) {
+  public Quiz findById(Long id) {
     return quizRepository.findById(id);
   }
