+++ week3/d5-mvc-jsp/d5_mvc_jsp_quiz/src/main/java/com/example/d5_mvc_jsp_quiz/service/UserService.java
@@ -22,7 +22,7 @@ public class UserService implements ObjectService<User> {
   }
 
   @Override
-  public User findById(int id) {
+  public User findById(Long id) {
     return userRepository.findById(id);
   }
